+++ src/app/blog/page.tsx
@@ -1,6 +1,8 @@
+import { GroupPreviewContent } from "@/components/group/_groupTypes";
 import GroupPage from "@/components/group/groupPage/_groupPage";
 import { generateCustomMetadata } from "@/content/Metadata";
 import { MirrorSynopsis, MirrorTitle } from "@/content/blog/mirrorsEdge";
+import { imageFolder } from "@/content/portfolio/Portfolio";
 import { Metadata } from "next";
 import { OpenGraph } from "next/dist/lib/metadata/types/opengraph-types";
 
@@ -17,6 +19,24 @@ export const metadata: Metadata = generateCustomMetadata({
         description,
     } as OpenGraph,
 });
+
+const Marta: GroupPreviewContent = {
+    title: "Marta's Fundraiser",
+    desc: `
+    My mom and I helped organize a community fundraiser in Bolton and Bury for Marta, a single mother with cancer who lacked basics like furniture and a clothes dryer. Thanks to the generosity of our communities, we were able to gather the funds needed to support her and her three young children during this challenging time.
+    `,
+    isDemo: false,
+    image: imageFolder + "marta.jpg",
+    url: preURL + "marta",
+}
+
+const crab: GroupPreviewContent = {
+    title: "Hobbies",
+    desc: "Sometimes I like to take a break from programming. I enjoy making 3d art, music and other random things.",
+    isDemo: false,
+    url: preURL + "hobbies",
+    image: imageFolder + "crab_rave.png",
+};
 
 export default function Blog() {
     return <GroupPage
@@ -48,6 +68,14 @@ export default function Blog() {
                             url: "blog/mirrors_edge",
                             image: `${preURL}mirrors_edge/1x1 partial logo.png`,
                         },
+                    ],
+                },
+                {
+                    title: "Extra things",
+                    subtitle: "",
+                    groups: [
+                        Marta,
+                        crab,
                     ],
                 },
             ]
